\name{OccDesign-package}
\alias{OccDesign-package}
\alias{OccDesign}
\docType{package}
\title{
\packageTitle{OccDesign}
}
\description{
\packageDescription{OccDesign}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OccDesign}
\packageIndices{OccDesign}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{OccDesign}

Maintainer: \packageMaintainer{OccDesign}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
